examples/custom_llm_gemini/pyproject.toml
[tool.poetry]
name = "llmproxy"
version = "0.1.0"
description = ""
authors = []
packages = [{include = "src"}]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
flask = "^3.0.3"
requests = "^2.32.3"
aiohttp = "^3.10.10"
python-dotenv = "^1.0.1"
flask-cors = "^5.0.0"
rich = "^13.9.2"
pydantic = "^2.9.2"
google-generativeai = "^0.8.3"
h11 = "^0.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# CLI command configuration
[tool.poetry.scripts]
llm = "src.llm:run_llm"
